Finalized name change and example
