+++ lib/spark_rest.dart
@@ -4,14 +4,12 @@
 library spark_rest;
 
 export 'src/spark_rest_base.dart';
-export 'src/server/actuator/application.dart';
 export 'src/server/actuator/endpoint.dart';
 export 'src/server/actuator/middleware.dart';
 export 'src/server/container/method.dart';
 export 'src/server/container/request.dart';
 export 'src/server/container/response.dart';
-export 'src/server/container/middleware_attach_type.dart';
+export 'src/server/foundation/application.dart';
+export 'src/server/foundation/context.dart';
 export 'src/server/router/method.dart';
 export 'src/server/router/uri.dart';
-export 'src/server/container/context.dart';
-export 'src/addon/plugin.dart';
