--- lib/src/abstract_server/interface/handlable.dart
@@ -1,5 +1,0 @@
-abstract class Handlable<U, V> {
-  
-  Future<V> onHandle(U param);
-
-}
