--- lib/src/abstract_server/interface/initializable.dart
@@ -1,7 +1,0 @@
-import 'package:spark_rest/src/abstract_server/foundation/context.dart';
-
-abstract class Initializable {
-
-  Future<void> onInit(Context context);
-
-}
