+++ lib/src/server/interface/handlable.dart
@@ -1,7 +1,5 @@
-/// An interface for all handlable components
-abstract class Handlable<R, P> {
-  /// Called every time an handling is needed.
-  ///
-  /// Do not call this function, it is called from SparkREST from you.
-  Future<R> onHandle(final P param);
+abstract class Handlable<U, V> {
+  
+  Future<V> onHandle(U param);
+
 }
