+++ lib/src/server/interface/initializable.dart
@@ -1,9 +1,7 @@
-import 'package:spark_rest/src/server/container/context.dart';
+import 'package:spark_rest/src/server/foundation/context.dart';
 
-/// An interface for all initializable components
 abstract class Initializable {
-  /// Called to init a component only one time.
-  ///
-  /// Do not call this function, it is called from SparkREST from you.
-  Future<void> onInit(final Context context);
+
+  Future<void> onInit(Context context);
+
 }
